+++ devicemanagement/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- Conanfile .libs specification
 ### Security
 
 ## [0.4.0] - 2022-07-07